requirements.txt
torch
torchvision
numpy
opencv-python-headless
tensorboard
tqdm
Pillow
